backend/requirements.txt
cohere>=5.0.0
qdrant-client==1.12.0
beautifulsoup4
requests
python-dotenv
fastapi
uvicorn
google-generativeai
pydantic
